Add a new flow for the send function
It now starts a QR scan window as first action, and by default at the
end of if shows the window where the user has to fill in amount and
recipient (with possible prefilling from QR).

If the QR scanned is a manta address, start a manta flow of with
confirmation and subsequent payment.

If `contact` or `address` or `quickSendAmount` are defined at
instantiation time, again shows the form with prefilling, with no QR
stage.